Refactor
Move all tests out of activities package
