+++ app/src/androidTest/java/com/dcrandroid/SeedBackupTest.kt
@@ -4,7 +4,7 @@
  * license that can be found in the LICENSE file.
  */
 
-package com.dcrandroid.activities
+package com.dcrandroid
 
 
 import android.view.View
@@ -25,6 +25,7 @@ import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
 import androidx.test.runner.AndroidJUnit4
 import com.dcrandroid.R
+import com.dcrandroid.activities.SplashScreenActivity
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
 import org.junit.Rule
