+++ app/src/androidTest/java/com/dcrandroid/Util.kt
@@ -4,7 +4,7 @@
  * license that can be found in the LICENSE file.
  */
 
-package com.dcrandroid.activities
+package com.dcrandroid
 
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.espresso.Espresso
